backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
google-cloud-storage>=2.18.0
google-cloud-firestore>=2.19.0
firebase-admin>=6.6.0
pydub>=0.25.1
numpy>=1.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
